New app: expense
The "expense" app will focus on managing the expenses of each user.
For now, only a very basic model has been defined.

An expense has an owner (the person who spent the money), a name
and an amount, and also one or possibly more recipients.

Ex: Tom can buy something only for himself, or for Alice and him,
or everyone in the house.
Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>