New app: roommate
The roommate app will manage the colocation roommates actions and views.

Warning: a differentiation must be made between a roommate and a user.
A user is simply used for login, authentication, acl, etc. The roommate
is the "buisness" user model, holding all the expenses, etc.

There is a OneToOne relationship bewteen the Roommate and the User,
and each time a User is created, a Roommate is autmagically created,
using the User.post_save signal.

The Roommate forms have been tweaked (using ModelForms) so that they
use the User forms.
Signed-off-by: Balthazar Rouberol <balthazar.rouberol@mapado.com>